GATEWAY-2946: Fix EventFactory to return a new Event for each call to each method.
