--- Product/Production/Common/CONNECTCoreLib/src/main/java/gov/hhs/fha/nhinc/event/BaseEventFactory.java
@@ -1,109 +1,0 @@
-/**
- * Copyright (c) 2012, United States Government, as represented by the Secretary of Health and Human Services.
- * All rights reserved.
- *
- * Redistribution and use in source and binary forms, with or without
- * modification, are permitted provided that the following conditions are met:
- *     * Redistributions of source code must retain the above
- *       copyright notice, this list of conditions and the following disclaimer.
- *     * Redistributions in binary form must reproduce the above copyright
- *       notice, this list of conditions and the following disclaimer in the documentation
- *       and/or other materials provided with the distribution.
- *     * Neither the name of the United States Government nor the
- *       names of its contributors may be used to endorse or promote products
- *       derived from this software without specific prior written permission.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
- * ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
- * WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
- * DISCLAIMED. IN NO EVENT SHALL THE UNITED STATES GOVERNMENT BE LIABLE FOR ANY
- * DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
- * (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
- * LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND
- * ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
- * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
- * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- */
-package gov.hhs.fha.nhinc.event;
-
-import java.util.HashMap;
-import java.util.Map;
-
-public class BaseEventFactory implements EventFactory {
-
-    private Map<String, Event> eventMap;
-
-    public BaseEventFactory() {
-        eventMap = new HashMap<String, Event>();
-    }
-
-    public void setEventMap(Map<String, Event> eventMap) {
-        this.eventMap = eventMap;
-    }
-
-    @Override
-    public Event createBeginOutboundMessage() {
-        return eventMap.get("beginOutboundMessage");
-    }
-
-    @Override
-    public Event createBeginOutboundProcessing() {
-        return eventMap.get("beginOutboundProcessing");
-    }
-
-    @Override
-    public Event createBeginNwhinInvocation() {
-        return eventMap.get("beginNwhinInvocation");
-    }
-
-    @Override
-    public Event createEndNwhinInvocation() {
-        return eventMap.get("endNwhinInvocation");
-    }
-
-    @Override
-    public Event createEndOutboundProcessing() {
-        return eventMap.get("endOutboundProcessing");
-    }
-
-    @Override
-    public Event createEndOutboundMessage() {
-        return eventMap.get("endOutboundMessage");
-    }
-
-    @Override
-    public Event createBeginInboundMessage() {
-        return eventMap.get("beginInboundMessage");
-    }
-
-    @Override
-    public Event createBeginInboundProcessing() {
-        return eventMap.get("beginInboundProcessing");
-    }
-
-    @Override
-    public Event createBeginAdapterDelegation() {
-        return eventMap.get("beginAdapterDelegation");
-    }
-
-    @Override
-    public Event createEndAdapterDelegation() {
-        return eventMap.get("endAdapterDelegation");
-    }
-
-    @Override
-    public Event createEndInboundProcessing() {
-        return eventMap.get("endInboundProcessing");
-    }
-
-    @Override
-    public Event createEndInboundMessage() {
-        return eventMap.get("endInboundMessage");
-    }
-
-    @Override
-    public Event createMessageProcessingFailed() {
-        return eventMap.get("messageProcessingFailed");
-    }
-
-}
