+++ Product/Production/Common/Properties/Dev/EventFactoryConfig.xml
@@ -10,75 +10,5 @@
     
     <alias alias="eventfactory" name="baseeventfactory"/>
 
-    <bean name="baseeventfactory" class="gov.hhs.fha.nhinc.event.BaseEventFactory">
-        <property name="eventMap">
-            <map>
-                <entry key="beginOutboundMessage">
-                    <bean
-                        class="gov.hhs.fha.nhinc.event.initiator.BeginOutboundMessageEvent"
-                        scope="prototype" />
-                </entry>
-                <entry key="beginOutboundProcessing">
-                    <bean
-                        class="gov.hhs.fha.nhinc.event.initiator.BeginOutboundProcessingEvent"
-                        scope="prototype" />
-                </entry>
-                <entry key="beginNwhinInvocation">
-                    <bean
-                        class="gov.hhs.fha.nhinc.event.initiator.BeginNwhinInvocationEvent"
-                        scope="prototype" />
-                </entry>
-                <entry key="endNwhinInvocation">
-                    <bean
-                        class="gov.hhs.fha.nhinc.event.initiator.EndNwhinInvocationEvent"
-                        scope="prototype" />
-                </entry>
-                <entry key="endOutboundProcessing">
-                    <bean
-                        class="gov.hhs.fha.nhinc.event.initiator.EndOutboundProcessingEvent"
-                        scope="prototype" />
-                </entry>
-                <entry key="endOutboundMessage">
-                    <bean
-                        class="gov.hhs.fha.nhinc.event.initiator.EndOutboundMessageEvent"
-                        scope="prototype" />
-                </entry>
-                <entry key="beginInboundMessage">
-                    <bean
-                        class="gov.hhs.fha.nhinc.event.responder.BeginInboundMessageEvent"
-                        scope="prototype" />
-                </entry>
-                <entry key="beginInboundProcessing">
-                    <bean
-                        class="gov.hhs.fha.nhinc.event.responder.BeginInboundProcessingEvent"
-                        scope="prototype" />
-                </entry>
-                <entry key="">
-                    <bean id="beginAdapterDelegation"
-                        class="gov.hhs.fha.nhinc.event.responder.BeginAdapterDelegationEvent"
-                        scope="prototype" />
-                </entry>
-                <entry key="endAdapterDelegation">
-                    <bean
-                        class="gov.hhs.fha.nhinc.event.responder.EndAdapterDelegationEvent"
-                        scope="prototype" />
-                </entry>
-                <entry key="endInboundProcessing">
-                    <bean
-                        class="gov.hhs.fha.nhinc.event.responder.EndInboundProcessingEvent"
-                        scope="prototype" />
-                </entry>
-                <entry key="endInboundMessage">
-                    <bean
-                        class="gov.hhs.fha.nhinc.event.responder.EndInboundMessageEvent"
-                        scope="prototype" />
-                </entry>
-                <entry key="messageProcessingFailed">
-                    <bean
-                        class="gov.hhs.fha.nhinc.event.error.MessageProcessingFailedEvent"
-                        scope="prototype" />
-                </entry>
-            </map>
-        </property>
-    </bean>
+    <bean name="baseeventfactory" class="gov.hhs.fha.nhinc.event.EventFactory"/>
 </beans>
